+++ scinoephile/audio/separation/demucs/__init__.py
@@ -1,6 +1,6 @@
 #  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
 #  and distributed under the terms of the BSD license. See the LICENSE file for details.
-"""Demucs vocal separation for transcription preprocessing.
+"""Demucs audio source separation.
 
 Package hierarchy (modules may import from any above):
 * cache
