+++ scinoephile/audio/transcription/mlx_audio/transcriber.py
@@ -13,8 +13,8 @@ from typing import TYPE_CHECKING
 
 import numpy as np
 
+from scinoephile.audio.separation import DemucsSeparator
 from scinoephile.audio.transcription.ctc_aligner import CtcAligner
-from scinoephile.audio.transcription.demucs import DemucsSeparator
 from scinoephile.audio.transcription.exceptions import (
     TranscriptionAlignmentError,
     TranscriptionAlignmentIncompleteError,
